+++ charts/jfrog-registry-operator/CHANGELOG.md
@@ -1,6 +1,16 @@
 # JFrog Secret Rotator Operator Chart Changelog
 All changes to this chart will be documented in this file.
 
+## [2.0.0] - May 19, 2025
+*** Important Changes ***
+* In the custom resource, the introduced `spec.generatedSecrets` configuration typically involves specifying: `secretName` – the name of the Secret to be generated, and `secretType` – the type of Secret to generate (e.g., Docker, Generic)
+* Scope : Scope can be anything (Optional)
+* Note: Currently spec.secretName is supported but going forward this will be deprecated soon.
+
+## [1.4.2] - Mar 26, 2025
+* Release of jfrog-registry-operator `1.4.2`
+* Added support for providing containerSecurityContext
+
 ## [1.4.1] - Sept 03, 2024
 * Release of jfrog-registry-operator `1.4.1`